requirements.txt
aiohttp>=3.7.4
google-auth-oauthlib>=0.4.1
google-auth>=1.30.0
Pillow>=8.0.0
torch>=1.7.0
termcolor>=1.1.0
numpy>=1.19.0